functions/genthumbnail/requirements.txt
requests                                                                                                                                                                  
minio
pillow
